+++ racetrack_client/racetrack_client/__init__.py
@@ -1,2 +1,2 @@
 name = "racetrack_client"
-__version__ = "2.2.1"
+__version__ = "2.3.0"