+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Plugins are distributed as ZIP files.
   They can be installed and uninstalled in a Dashboard's Administration page.
+  See [plugins.md](development/plugins.md).
 
 ## [2.2.1] - 2022-08-25
 ### Changed